added:: WDS preset for tailwind to make it resuable.
